+++ variables.tf
@@ -111,6 +111,12 @@ variable "desired_capacity" {
   default     = null
 }
 
+variable "desired_capacity_type" {
+  type        = string
+  description = "(Optional) The unit of measurement for the value specified for `desired_capacity`. Supported for attribute-based instance type selection only. Valid values: \"units\", \"vcpu\", \"memory-mib\"."
+  default     = null
+}
+
 variable "force_delete" {
   type        = bool
   description = "(Optional) Allows deleting the Auto Scaling Group without waiting for all instances in the pool to terminate. You can force an Auto Scaling Group to delete even if it's in the process of scaling a resource. Normally, Terraform drains all the instances before deleting the group. This bypasses that behavior and potentially leaves resources dangling."